requirements.txt
torch
langchain
pillow
OpenAI
dotenv
gradio
google-search-results
gradio_multimodalchatbot